+++ config/defaults.php
@@ -13,26 +13,15 @@
 
 namespace BrightNucleus\CustomContent;
 
-use BrightNucleus\CustomContent\CustomPostType\Argument as CPTArgument;
-use BrightNucleus\CustomContent\CustomPostType\Label as CPTLabel;
-use BrightNucleus\CustomContent\CustomPostType\Link as CPTLink;
-use BrightNucleus\CustomContent\CustomPostType\MenuPosition as CPTMenuPosition;
-use BrightNucleus\CustomContent\CustomPostType\Message as CPTMessage;
-use BrightNucleus\CustomContent\CustomPostType\Name as CPTName;
-use BrightNucleus\CustomContent\CustomTaxonomy\Argument as TaxArgument;
-use BrightNucleus\CustomContent\CustomTaxonomy\Capability as TaxCapability;
-use BrightNucleus\CustomContent\CustomTaxonomy\Label as TaxLabel;
-use BrightNucleus\CustomContent\CustomTaxonomy\Name as TaxName;
-
 /*
  * Default upper-case and lower case versions of single and plural name for
  * custom post type.
  */
 $cpt_default_names = [
-	CPTName::SINGULAR_NAME_UC => _x( 'Element', 'post type upper case singular name', 'bn-custom-content' ),
-	CPTName::SINGULAR_NAME_LC => _x( 'element', 'post type lower case singular name', 'bn-custom-content' ),
-	CPTName::PLURAL_NAME_UC   => _x( 'Elements', 'post type upper case plural name', 'bn-custom-content' ),
-	CPTName::PLURAL_NAME_LC   => _x( 'elements', 'post type lower case plural name', 'bn-custom-content' ),
+	CustomPostType\Name::SINGULAR_NAME_UC => _x( 'Element', 'post type upper case singular name', 'bn-custom-content' ),
+	CustomPostType\Name::SINGULAR_NAME_LC => _x( 'element', 'post type lower case singular name', 'bn-custom-content' ),
+	CustomPostType\Name::PLURAL_NAME_UC   => _x( 'Elements', 'post type upper case plural name', 'bn-custom-content' ),
+	CustomPostType\Name::PLURAL_NAME_LC   => _x( 'elements', 'post type lower case plural name', 'bn-custom-content' ),
 ];
 
 /*
@@ -40,53 +29,53 @@ $cpt_default_names = [
  */
 $cpt_default_labels = [
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	CPTLabel::NAME                  => __( '%3$s', 'bn-custom-content' ),
+	CustomPostType\Label::NAME                  => _x( '%3$s', 'cpt general name', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	CPTLabel::SINGULAR_NAME         => __( '%1$s', 'bn-custom-content' ),
+	CustomPostType\Label::SINGULAR_NAME         => _x( '%1$s', 'cpt singular name', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	CPTLabel::ADD_NEW               => _x( 'Add New', 'element', 'bn-custom-content' ),
+	CustomPostType\Label::ADD_NEW               => _x( 'Add New', 'element', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	CPTLabel::ADD_NEW_ITEM          => __( 'Add New %1$s', 'bn-custom-content' ),
+	CustomPostType\Label::ADD_NEW_ITEM          => __( 'Add New %1$s', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	CPTLabel::EDIT_ITEM             => __( 'Edit %1$s', 'bn-custom-content' ),
+	CustomPostType\Label::EDIT_ITEM             => __( 'Edit %1$s', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	CPTLabel::NEW_ITEM              => __( 'New %1$s', 'bn-custom-content' ),
+	CustomPostType\Label::NEW_ITEM              => __( 'New %1$s', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	CPTLabel::VIEW_ITEM             => __( 'View %1$s', 'bn-custom-content' ),
+	CustomPostType\Label::VIEW_ITEM             => __( 'View %1$s', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	CPTLabel::SEARCH_ITEMS          => __( 'Search %3$s', 'bn-custom-content' ),
+	CustomPostType\Label::SEARCH_ITEMS          => __( 'Search %3$s', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	CPTLabel::NOT_FOUND             => __( 'No %4$s found.', 'bn-custom-content' ),
+	CustomPostType\Label::NOT_FOUND             => __( 'No %4$s found.', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	CPTLabel::NOT_FOUND_IN_TRASH    => __( 'No %4$s found in Trash.', 'bn-custom-content' ),
+	CustomPostType\Label::NOT_FOUND_IN_TRASH    => __( 'No %4$s found in Trash.', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	CPTLabel::PARENT_ITEM_COLON     => __( 'Parent %1$s:', 'bn-custom-content' ),
+	CustomPostType\Label::PARENT_ITEM_COLON     => __( 'Parent %1$s:', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	CPTLabel::ALL_ITEMS             => __( 'All %3$s', 'bn-custom-content' ),
+	CustomPostType\Label::ALL_ITEMS             => __( 'All %3$s', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	CPTLabel::ARCHIVES              => __( '%1$s Archives', 'bn-custom-content' ),
+	CustomPostType\Label::ARCHIVES              => __( '%1$s Archives', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	CPTLabel::INSERT_INTO_ITEM      => __( 'Insert into %2$s', 'bn-custom-content' ),
+	CustomPostType\Label::INSERT_INTO_ITEM      => __( 'Insert into %2$s', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	CPTLabel::UPLOADED_TO_THIS_ITEM => __( 'Uploaded to this %2$s', 'bn-custom-content' ),
+	CustomPostType\Label::UPLOADED_TO_THIS_ITEM => __( 'Uploaded to this %2$s', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	CPTLabel::FEATURED_IMAGE        => __( 'Featured Image', 'bn-custom-content' ),
+	CustomPostType\Label::FEATURED_IMAGE        => __( 'Featured Image', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	CPTLabel::SET_FEATURED_IMAGE    => __( 'Set featured image', 'bn-custom-content' ),
+	CustomPostType\Label::SET_FEATURED_IMAGE    => __( 'Set featured image', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	CPTLabel::REMOVE_FEATURED_IMAGE => __( 'Remove featured image', 'bn-custom-content' ),
+	CustomPostType\Label::REMOVE_FEATURED_IMAGE => __( 'Remove featured image', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	CPTLabel::USE_FEATURED_IMAGE    => __( 'Use as featured image', 'bn-custom-content' ),
+	CustomPostType\Label::USE_FEATURED_IMAGE    => __( 'Use as featured image', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	CPTLabel::MENU_NAME             => _x( '%3$s', 'admin menu', 'bn-custom-content' ),
+	CustomPostType\Label::MENU_NAME             => _x( '%3$s', 'admin menu', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	CPTLabel::FILTER_ITEMS_LIST     => __( 'Filter %4$s list', 'bn-custom-content' ),
+	CustomPostType\Label::FILTER_ITEMS_LIST     => __( 'Filter %4$s list', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	CPTLabel::ITEMS_LIST_NAVIGATION => __( '%3$s list navigation', 'bn-custom-content' ),
+	CustomPostType\Label::ITEMS_LIST_NAVIGATION => __( '%3$s list navigation', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	CPTLabel::ITEMS_LIST            => __( '%3$s list', 'bn-custom-content' ),
+	CustomPostType\Label::ITEMS_LIST            => __( '%3$s list', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	CPTLabel::NAME_ADMIN_BAR        => _x( '%1$s', 'add new on admin bar', 'bn-custom-content' ),
+	CustomPostType\Label::NAME_ADMIN_BAR        => _x( '%1$s', 'add new on admin bar', 'bn-custom-content' ),
 ];
 
 /*
@@ -94,27 +83,27 @@ $cpt_default_labels = [
  */
 $cpt_default_messages = [
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	CPTMessage::UNUSED                    => '', // Unused. Messages start at index 1.
+	CustomPostType\Message::UNUSED                    => '', // Unused. Messages start at index 1.
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	CPTMessage::ELEMENT_UPDATED_WITH_LINK => __( '%1$s updated.', 'bn-custom-content' ), // View Element link will be appended.
+	CustomPostType\Message::ELEMENT_UPDATED_WITH_LINK => __( '%1$s updated.', 'bn-custom-content' ), // View Element link will be appended.
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	CPTMessage::CUSTOM_FIELD_UPDATED      => __( 'Custom field updated.', 'bn-custom-content' ),
+	CustomPostType\Message::CUSTOM_FIELD_UPDATED      => __( 'Custom field updated.', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	CPTMessage::CUSTOM_FIELD_DELETED      => __( 'Custom field deleted.', 'bn-custom-content' ),
+	CustomPostType\Message::CUSTOM_FIELD_DELETED      => __( 'Custom field deleted.', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	CPTMessage::ELEMENT_UPDATED           => __( '%1$s updated.', 'bn-custom-content' ),
+	CustomPostType\Message::ELEMENT_UPDATED           => __( '%1$s updated.', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural, %5$s date and time of the revision. */
-	CPTMessage::REVISION_RESTORED         => __( '%1$s restored to revision from %5$s.', 'bn-custom-content' ),
+	CustomPostType\Message::REVISION_RESTORED         => __( '%1$s restored to revision from %5$s.', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	CPTMessage::ELEMENT_PUBLISHED         => __( '%1$s published.', 'bn-custom-content' ), // View Element link will be appended.
+	CustomPostType\Message::ELEMENT_PUBLISHED         => __( '%1$s published.', 'bn-custom-content' ), // View Element link will be appended.
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	CPTMessage::ELEMENT_SAVED             => __( '%1$s saved.', 'bn-custom-content' ),
+	CustomPostType\Message::ELEMENT_SAVED             => __( '%1$s saved.', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	CPTMessage::ELEMENT_SUBMITTED         => __( '%1$s submitted.', 'bn-custom-content' ), // Preview Element link will be appended.
+	CustomPostType\Message::ELEMENT_SUBMITTED         => __( '%1$s submitted.', 'bn-custom-content' ), // Preview Element link will be appended.
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural, %5$s scheduled date. */
-	CPTMessage::ELEMENT_SCHEDULED         => __( '%1$s scheduled for: %5$s.', 'bn-custom-content' ), // Scheduled Page link will be appended.
+	CustomPostType\Message::ELEMENT_SCHEDULED         => __( '%1$s scheduled for: %5$s.', 'bn-custom-content' ), // Scheduled Page link will be appended.
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	CPTMessage::ELEMENT_DRAFT_UPDATED     => __( '%1$s draft updated.', 'bn-custom-content' ),
+	CustomPostType\Message::ELEMENT_DRAFT_UPDATED     => __( '%1$s draft updated.', 'bn-custom-content' ),
 	// Preview Element link will be appended.
 ];
 
@@ -123,9 +112,9 @@ $cpt_default_messages = [
  */
 $cpt_default_links = [
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	CPTLink::VIEW_LINK    => __( 'View %2$s', 'bn-custom-content' ),
+	CustomPostType\Link::VIEW_LINK    => __( 'View %2$s', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	CPTLink::PREVIEW_LINK => __( 'Preview %2$s', 'bn-custom-content' ),
+	CustomPostType\Link::PREVIEW_LINK => __( 'Preview %2$s', 'bn-custom-content' ),
 ];
 
 /*
@@ -141,14 +130,14 @@ $cpt_default_supports = [
  */
 $cpt_defaults = [
 	// WordPress CPT registration arguments.
-	CPTArgument::LABELS        => $cpt_default_labels,
-	CPTArgument::IS_PUBLIC     => true,
-	CPTArgument::MENU_POSITION => CPTMenuPosition::BELOW_POSTS,
-	CPTArgument::SUPPORTS      => $cpt_default_supports,
+	CustomPostType\Argument::LABELS        => $cpt_default_labels,
+	CustomPostType\Argument::IS_PUBLIC     => true,
+	CustomPostType\Argument::MENU_POSITION => CustomPostType\MenuPosition::BELOW_POSTS,
+	CustomPostType\Argument::SUPPORTS      => $cpt_default_supports,
 	// Bright Nucleus Custom Post Type component arguments.
-	CPTArgument::NAMES         => $cpt_default_names,
-	CPTArgument::MESSAGES      => $cpt_default_messages,
-	CPTArgument::LINKS         => $cpt_default_links,
+	CustomPostType\Argument::NAMES         => $cpt_default_names,
+	CustomPostType\Argument::MESSAGES      => $cpt_default_messages,
+	CustomPostType\Argument::LINKS         => $cpt_default_links,
 ];
 
 /*
@@ -156,10 +145,10 @@ $cpt_defaults = [
  * custom post type.
  */
 $tax_default_names = [
-	TaxName::SINGULAR_NAME_UC => _x( 'Taxonomy', 'taxonomy upper case singular name', 'bn-custom-content' ),
-	TaxName::SINGULAR_NAME_LC => _x( 'taxonomy', 'taxonomy lower case singular name', 'bn-custom-content' ),
-	TaxName::PLURAL_NAME_UC   => _x( 'Taxonomies', 'taxonomy upper case plural name', 'bn-custom-content' ),
-	TaxName::PLURAL_NAME_LC   => _x( 'taxonomies', 'taxonomy lower case plural name', 'bn-custom-content' ),
+	CustomTaxonomy\Name::SINGULAR_NAME_UC => _x( 'Taxonomy', 'taxonomy upper case singular name', 'bn-custom-content' ),
+	CustomTaxonomy\Name::SINGULAR_NAME_LC => _x( 'taxonomy', 'taxonomy lower case singular name', 'bn-custom-content' ),
+	CustomTaxonomy\Name::PLURAL_NAME_UC   => _x( 'Taxonomies', 'taxonomy upper case plural name', 'bn-custom-content' ),
+	CustomTaxonomy\Name::PLURAL_NAME_LC   => _x( 'taxonomies', 'taxonomy lower case plural name', 'bn-custom-content' ),
 ];
 
 /*
@@ -167,49 +156,49 @@ $tax_default_names = [
  */
 $tax_default_labels = [
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	TaxLabel::NAME                       => _x( '%3$s', 'taxonomy general name', 'bn-custom-content' ),
+	CustomTaxonomy\Label::NAME                       => _x( '%3$s', 'taxonomy general name', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	TaxLabel::SINGULAR_NAME              => _x( '%1$s', 'taxonomy singular name', 'bn-custom-content' ),
+	CustomTaxonomy\Label::SINGULAR_NAME              => _x( '%1$s', 'taxonomy singular name', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	TaxLabel::MENU_NAME                  => _x( '%3$s', 'taxonomy menu name', 'bn-custom-content' ),
+	CustomTaxonomy\Label::MENU_NAME                  => _x( '%3$s', 'taxonomy menu name', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	TaxLabel::ALL_ITEMS                  => __( 'All %3$s', 'bn-custom-content' ),
+	CustomTaxonomy\Label::ALL_ITEMS                  => __( 'All %3$s', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	TaxLabel::EDIT_ITEM                  => __( 'Edit %1$s', 'bn-custom-content' ),
+	CustomTaxonomy\Label::EDIT_ITEM                  => __( 'Edit %1$s', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	TaxLabel::VIEW_ITEM                  => __( 'View %1$s', 'bn-custom-content' ),
+	CustomTaxonomy\Label::VIEW_ITEM                  => __( 'View %1$s', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	TaxLabel::UPDATE_ITEM                => __( 'Update %1$s', 'bn-custom-content' ),
+	CustomTaxonomy\Label::UPDATE_ITEM                => __( 'Update %1$s', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	TaxLabel::ADD_NEW_ITEM               => __( 'Add New %1$s', 'bn-custom-content' ),
+	CustomTaxonomy\Label::ADD_NEW_ITEM               => __( 'Add New %1$s', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	TaxLabel::NEW_ITEM_NAME              => __( 'New %1$s Name', 'bn-custom-content' ),
+	CustomTaxonomy\Label::NEW_ITEM_NAME              => __( 'New %1$s Name', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	TaxLabel::PARENT_ITEM                => __( 'Parent %1$s', 'bn-custom-content' ),
+	CustomTaxonomy\Label::PARENT_ITEM                => __( 'Parent %1$s', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	TaxLabel::PARENT_ITEM_COLON          => __( 'Parent %1$s:', 'bn-custom-content' ),
+	CustomTaxonomy\Label::PARENT_ITEM_COLON          => __( 'Parent %1$s:', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	TaxLabel::SEARCH_ITEMS               => __( 'Search %3$s', 'bn-custom-content' ),
+	CustomTaxonomy\Label::SEARCH_ITEMS               => __( 'Search %3$s', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	TaxLabel::POPULAR_ITEMS              => __( 'Popular %3$s', 'bn-custom-content' ),
+	CustomTaxonomy\Label::POPULAR_ITEMS              => __( 'Popular %3$s', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	TaxLabel::SEPARATE_ITEMS_WITH_COMMAS => __( 'Separate %4$s with commas', 'bn-custom-content' ),
+	CustomTaxonomy\Label::SEPARATE_ITEMS_WITH_COMMAS => __( 'Separate %4$s with commas', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	TaxLabel::ADD_OR_REMOVE_ITEMS        => __( 'Add or remove %4$s', 'bn-custom-content' ),
+	CustomTaxonomy\Label::ADD_OR_REMOVE_ITEMS        => __( 'Add or remove %4$s', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	TaxLabel::CHOOSE_FROM_MOST_USED      => __( 'Choose from the most used %4$s', 'bn-custom-content' ),
+	CustomTaxonomy\Label::CHOOSE_FROM_MOST_USED      => __( 'Choose from the most used %4$s', 'bn-custom-content' ),
 	/* Translators: %1$s uc singular, %2$s lc singular, %3$s uc plural, %4$s lc plural. */
-	TaxLabel::NOT_FOUND                  => __( 'No %4$s found.', 'bn-custom-content' ),
+	CustomTaxonomy\Label::NOT_FOUND                  => __( 'No %4$s found.', 'bn-custom-content' ),
 ];
 
 /*
  * Default capabilities of a custom taxonomy.
  */
 $tax_default_capabilities = [
-	TaxCapability::MANAGE_TERMS => 'manage_categories',
-	TaxCapability::EDIT_TERMS   => 'manage_categories',
-	TaxCapability::DELETE_TERMS => 'manage_categories',
-	TaxCapability::ASSIGN_TERMS => 'edit_posts',
+	CustomTaxonomy\Capability::MANAGE_TERMS => 'manage_categories',
+	CustomTaxonomy\Capability::EDIT_TERMS   => 'manage_categories',
+	CustomTaxonomy\Capability::DELETE_TERMS => 'manage_categories',
+	CustomTaxonomy\Capability::ASSIGN_TERMS => 'edit_posts',
 ];
 
 /*
@@ -217,11 +206,11 @@ $tax_default_capabilities = [
  */
 $tax_defaults = [
 	// WordPress custom taxonomy registration arguments.
-	TaxArgument::LABELS       => $tax_default_labels,
-	TaxArgument::CAPABILITIES => $tax_default_capabilities,
+	CustomTaxonomy\Argument::LABELS       => $tax_default_labels,
+	CustomTaxonomy\Argument::CAPABILITIES => $tax_default_capabilities,
 	// Bright Nucleus Custom Taxonomy component arguments.
-	TaxArgument::NAMES        => $tax_default_names,
-	TaxArgument::POST_TYPES   => [ ],
+	CustomTaxonomy\Argument::NAMES        => $tax_default_names,
+	CustomTaxonomy\Argument::POST_TYPES   => [ ],
 ];
 
 /*