+++ src/CustomPostType/ArgumentCollection.php
@@ -107,9 +107,10 @@ class ArgumentCollection extends ArrayCollection {
 				case Message::ELEMENT_SCHEDULED:
 					$post = get_post();
 
-					$args[4] = date_i18n( __( 'M j, Y @ G:i',
-						'bn-custom-content' ),
-						strtotime( $post->post_date ) );
+					$args[4] = date_i18n(
+						__( 'M j, Y @ G:i', 'bn-custom-content' ),
+						strtotime( $post->post_date )
+					);
 
 					break;