+++ src/CustomTaxonomy.php
@@ -27,9 +27,19 @@ use Doctrine\Common\Collections\Collection;
 class CustomTaxonomy extends AbstractContentType {
 
 	/*
-	 * Key that stores default values.
+	 * Reserved terms that cannot be used as custom taxonomy slug.
 	 */
-	const DEFAULTS = '_tax_defaults_';
+	const RESERVED_TERMS = [
+		'post',
+		'page',
+		'attachment',
+		'revision',
+		'nav_menu_item',
+		'action',
+		'author',
+		'order',
+		'theme',
+	];
 
 	/**
 	 * Prepare the arguments and return a Collection of arguments.
@@ -64,97 +74,5 @@ class CustomTaxonomy extends AbstractContentType {
 			$argsCollection->get( Argument::POST_TYPES ),
 			$argsCollection->toArray()
 		);
-	}
-
-	/**
-	 * Check whether a slug is a reserved term.
-	 *
-	 * @since 0.1.0
-	 *
-	 * @param string $slug Slug to check.
-	 * @return bool Whether the term is reserved.
-	 */
-	protected function isReservedTerm( $slug ) {
-		return in_array( $slug, [
-			'attachment',
-			'attachment_id',
-			'author',
-			'author_name',
-			'calendar',
-			'cat',
-			'category',
-			'category__and',
-			'category__in',
-			'category__not_in',
-			'category_name',
-			'comments_per_page',
-			'comments_popup',
-			'customize_messenger_channel',
-			'customized',
-			'cpage',
-			'day',
-			'debug',
-			'error',
-			'exact',
-			'feed',
-			'fields',
-			'hour',
-			'link_category',
-			'm',
-			'minute',
-			'monthnum',
-			'more',
-			'name',
-			'nav_menu',
-			'nonce',
-			'nopaging',
-			'offset',
-			'order',
-			'orderby',
-			'p',
-			'page',
-			'page_id',
-			'paged',
-			'pagename',
-			'pb',
-			'perm',
-			'post',
-			'post__in',
-			'post__not_in',
-			'post_format',
-			'post_mime_type',
-			'post_status',
-			'post_tag',
-			'post_type',
-			'posts',
-			'posts_per_archive_page',
-			'posts_per_page',
-			'preview',
-			'robots',
-			's',
-			'search',
-			'second',
-			'sentence',
-			'showposts',
-			'static',
-			'subpost',
-			'subpost_id',
-			'tag',
-			'tag__and',
-			'tag__in',
-			'tag__not_in',
-			'tag_id',
-			'tag_slug__and',
-			'tag_slug__in',
-			'taxonomy',
-			'tb',
-			'term',
-			'theme',
-			'type',
-			'w',
-			'withcomments',
-			'withoutcomments',
-			'year',
-		], true );
 	}
 }