+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.1.2] - 2016-08-15
+### Added
+- German translation.
+
+### Changed
+- Used custom constructors for exceptions.
+- Use `brightnucleus/localization` to make localization of Config strings work.
+
 ## [0.1.1] - 2016-08-11
 ### Fixed
 - Fix text domain.
@@ -11,5 +19,6 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Initial release to GitHub.
 
+[0.1.2]: https://github.com/brightnucleus/custom-content/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/brightnucleus/custom-content/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/brightnucleus/custom-content/compare/v0.0.0...v0.1.0